fix endpoint tests
